+++ config/dbc.go
@@ -1,7 +1,0 @@
-// Author: abbeymart | Abi Akindele | Created: 2018-12-05 | Updated: 2018-12-05
-// Company: mConnect.biz | License: MIT
-// Description: work: bookmark socket.io
-
-package config
-
-// Database connection package, including common data functions