+++ shared_template_functions.go
@@ -1,3 +1,0 @@
-// Shared Template functions: functions by template groups
-package utils
-