+++ stat_test.go
@@ -1,0 +1,5 @@
+// @Author: abbeymart | Abi Akindele | @Created: 2020-12-05 | @Updated: 2020-12-05
+// @Company: mConnect.biz | @License: MIT
+// @Description: go: mConnect
+
+package mcutils